+++ mytimer/functions.py
@@ -13,6 +13,7 @@ from mytimer.params import FACES_LIST_EXAMPLE_MESSAGE, TIME_PRINT_TEMPLATE
 from mytimer.params import DEFAULT_PARAMS, PROGRAMS_DEFAULTS, BREAKS_DEFAULTS
 from mytimer.params import NEXT_PROGRAM_MESSAGE, END_ROUND_MESSAGE
 from mytimer.params import KEEP_ON_MESSAGE, SET_ON_MESSAGE
+from mytimer.params import KEEP_ON_MAX
 from art import tprint
 
 
@@ -129,9 +130,6 @@ def load_params(args, program=None, is_break=False):
             else:
                 if not args.program:
                     params[item] = getattr(args, item)
-    if not args.countdown:
-        if check_null_time(args) and not args.program:
-            params["hour"] = 100000000
     return params
 
 
@@ -463,7 +461,7 @@ def keep_on_timer(params):
     :type params: dict
     :return: None
     """
-    params["hour"] = 10000000
+    params["hour"] = KEEP_ON_MAX
     params["message"] += KEEP_ON_MESSAGE
     if params["sign"] in ["+", ""]:
         params["sign"] = "-"
@@ -499,6 +497,28 @@ def update_set_on_params(params):
     return params
 
 
+def select_timer_func(args, params):
+    """
+    Select timer function.
+
+    :param args: input arguments
+    :type args: argparse.Namespace
+    :param params: timer params
+    :type params: dict
+    :return: timer function, timer params
+    """
+    timer_func = countdown_timer
+    if args.countup:
+        timer_func = countup_timer
+    if args.countdown:
+        timer_func = countdown_timer
+    else:
+        if check_null_time(args) and not args.program:
+            params["hour"] = KEEP_ON_MAX
+            timer_func = countup_timer
+    return timer_func, params
+
+
 def run_timer(args):
     """
     Run timer.
@@ -508,14 +528,9 @@ def run_timer(args):
     :return: None
     """
     params = load_params(args)
-    timer_func = countup_timer
+    timer_func, params = select_timer_func(args, params)
     if args.set_on:
-        timer_func = countdown_timer
         params = update_set_on_params(params)
-    if args.countdown:
-        timer_func = countdown_timer
-    if args.countup:
-        timer_func = countup_timer
     if args.version:
         print(MY_TIMER_VERSION)
     elif args.faces_list: