+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 	1. `puer-tea`
 	2. `purple-tea`
 - `--set-on` argument
+### Changed
+- Default mode changed from `count-up` to `countdown`
 ## [1.5] - 2024-08-12
 ### Added
 - 2 new programs