+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 	1. `mate`
 	2. `french-press`
 - 3 new tones
+### Changed
+- `README.md` updated
 ## [1.6] - 2024-08-30
 ### Added
 - 2 new programs